refactor: move files inside src
