--- output.txt
@@ -1,1 +1,0 @@
-
