+++ src/functions/argumentsCheckFn.js
