+++ src/streams/MyStdinStream.js
