+++ src/streams/myWriteStream.js
